requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets>=12.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg>=0.29.0
pgvector>=0.3.0

# Pydantic (upgraded for deepgram-sdk 6.x compatibility)
pydantic>=2.10.0
pydantic-settings>=2.6.0

# HTTP client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
loguru==0.7.3
tiktoken>=0.7.0

# Pipecat / Realtime pipeline
pipecat-ai==0.0.108
cartesia>=3.0.2

# Deepgram STT/TTS
deepgram-sdk>=6.0.1,<7
websockets>=14.1

# Audio/Math
numpy>=1.26.0
scipy==1.12.0
sounddevice>=0.5.0
PyAudio>=0.2.14

# Video processing (optional, for VMD upload)
opencv-python>=4.9.0

# OpenAI SDK (for vision LLM)
openai>=1.12.0

# Cerebras SDK (for direct Cerebras API access)
cerebras-cloud-sdk>=0.0.1

# Crypto (for Ed25519 if needed)
cryptography==42.0.0

# ========== 本地 Embedding ==========
sentence-transformers>=2.2.0
torch>=2.0.0  # sentence-transformers 依赖

# ========== 定时任务调度 ==========
apscheduler>=3.10.0
